Properly merged hazel-feature-branch into dev. Home page update
