+++ frontend/game/favicon.ico
