+++ frontend/game/html5game/splash.png
