+++ frontend/game/options.ini
@@ -1,0 +1,6 @@
+[HTML5]
+MajorVersion=1
+MinorVersion=0
+BuildVersion=0
+RevisionVersion=0
+DisplayName="2800Ladder"
