+++ frontend/views/autoRoute/home.ejs
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.5/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-SgOJa3DmI69IUzQ2PVdRZhwQ+dy64/BUtbMJw1MZ8t5HZApcHrRKUc4W0kG879m7" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.5/dist/js/bootstrap.bundle.min.js" integrity="sha384-k6d4wzSIapyDyv1kpU366/PK5hCdSbCRGRCMv+eplOQJWyd1fbcAu9OCUj5zNLiq" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="./css/main.css">
+    <title>Document</title>
+</head>
+<body>
+    <div class="card text-center text-bg-secondary m-2">
+        <div class="card-body">   
+            <div class="resizer">
+                <iframe class="resized" src="./game/index.html" title="Iframe Example"></iframe> 
+            </div>
+        </div>
+    </div>
+
+    <div class="card text-center text-bg-secondary m-2 p-2">
+
+        <!-- Create new task button -->
+        <div class="d-grid gap-2 d-md-flex justify-content-md-start mb-2">
+            <button class="btn btn-primary me-md-2" type="button">Create New Task</button>
+        </div>
+
+        <!-- Scroll With Cards inside -->
+        <div data-bs-spy="scroll" data-bs-target="#navbar-example2" data-bs-root-margin="0px 0px -40%" data-bs-smooth-scroll="true" class="scrollspy-example bg-body-tertiary p-3 rounded-2" tabindex="0" style="overflow-y: scroll; height: 300px;">
+            <div class="card text-bg-dark text-start mb-2">
+                <div class="card-body">
+                    <div class="row">
+                        <div class="col">
+                            <h1>Task Name</h1>
+                            <p>Short Description</p>
+                        </div>
+                        <div class="col d-grid gap-2">
+                            <button class="btn btn-primary" type="button">Complete Task</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div class="card text-bg-dark text-start mb-2">
+                <div class="card-body">
+                    <div class="row">
+                        <div class="col">
+                            <h1>Task Name</h1>
+                            <p>Short Description</p>
+                        </div>
+                        <div class="col d-grid gap-2">
+                            <button class="btn btn-primary" type="button">Complete Task</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>  
+    </div>
+
+    <!-- Footer -->
+    <div class="card text-center text-bg-secondary m-2 p-2">
+        <div class="row">
+            <div class="col">
+                <button class="btn btn-primary" type="button">Diary</button>
+            </div>
+            <div class="col">
+                <button class="btn btn-primary" type="button">Ladder</button>
+            </div>
+            <div class="col">
+                <button class="btn btn-primary" type="button">Profile/Stats</button>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
