--- index.js
@@ -1,20 +1,0 @@
-const express = require('express')
-const app = express()
-const port = 8100
-
-
-app.use(express.static('public'))
-
-app.get('/', (req, res) => {
-    res.sendFile(__dirname + '/public/html/main.html')
-})
-app.get('/game', (req, res) => {
-  res.sendFile(__dirname + '/public/game/index.html')
-})
-app.get('/about', (req, res) => {
-  res.sendFile(__dirname + '/public/html/about.html')
-})
-
-app.listen(port, () => {
-  console.log(`Example app listening on port ${port}`)
-})
